+++ lib/redmine_workload/hooks/after_plugins_loaded_hook.rb
@@ -4,8 +4,6 @@ module RedmineWorkload
   module Hooks
     class AfterPluginsLoadedHook < Redmine::Hook::Listener
       def after_plugins_loaded(_context = {})
-        return unless Gem::Version.new(Rails.version) >= Gem::Version.new('6.0')
-
         patch = RedmineWorkload::Extensions::UserPatch
         klass = User
         klass.prepend patch unless klass.included_modules.include?(patch)